@@ -1,8 +1,8 @@
 package com.example.coolcalculator
 
 import android.os.Bundle
-import android.support.design.widget.Snackbar
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity
+
 
 import kotlinx.android.synthetic.main.activity_graph.*
 
